move mobile auth related code to the mobile_app Django app
