+++ engine/apps/api/views/apns_device.py
@@ -1,6 +1,7 @@
 from push_notifications.api.rest_framework import APNSDeviceAuthorizedViewSet
 
-from apps.auth_token.auth import MobileAppAuthTokenAuthentication, PluginAuthentication
+from apps.auth_token.auth import PluginAuthentication
+from apps.mobile_app.auth import MobileAppAuthTokenAuthentication
 
 
 class APNSDeviceAuthorizedViewSet(APNSDeviceAuthorizedViewSet):
