+++ engine/apps/auth_token/constants.py
@@ -8,5 +8,3 @@ SLACK_AUTH_TOKEN_NAME = "slack_login_token"
 
 SCHEDULE_EXPORT_TOKEN_NAME = "token"
 SCHEDULE_EXPORT_TOKEN_CHARACTER_LENGTH = 32
-
-MOBILE_APP_AUTH_VERIFICATION_TOKEN_TIMEOUT_SECONDS = 60
